webapp/requirements-webapp.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
